requirements.txt
numpy==1.25.0
PyOpenGL==3.1.7
